Add more unit tests

+++ src/main/java/com/abn/recipes/repository/RecipesRepository.java
@@ -3,5 +3,5 @@ package com.abn.recipes.repository;
 import com.abn.recipes.model.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipesRepository extends JpaRepository<Recipe, Integer> {
+public interface RecipesRepository extends JpaRepository<Recipe, Long> {
 }
